Move _callOfFunction out and make general purpose
